Add files via upload
This Java Swing application calculates the engineering cutoff based on student marks in Chemistry, Physics, and Mathematics. Users can input their marks and ID to compute and display the cutoff score along with a history of past calculations.

copy paste the code in your eclipse ide and run it. prerequisites : install window builder in your ide